Refactor: auth 서버 디렉토리 정리
