--- auth/src/main/java/org/bookwoori/auth/controller/AuthController.java
@@ -1,25 +1,0 @@
-package org.bookwoori.auth.controller;
-
-import jakarta.validation.Valid;
-import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-import org.bookwoori.auth.dto.LoginResponseDto;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@Slf4j
-@RequiredArgsConstructor
-@RequestMapping("/auth")
-@RestController
-public class AuthController {
-
-    @GetMapping("/success")
-    public ResponseEntity<?> loginSuccess(@Valid LoginResponseDto loginResponseDto) {
-        return ResponseEntity.ok(loginResponseDto);
-    }
-
-
-}
-
