+++ auth/src/main/java/org/bookwoori/auth/domain/member/dto/LoginResponseDto.java
@@ -1,0 +1,7 @@
+package org.bookwoori.auth.domain.member.dto;
+
+import jakarta.validation.constraints.NotBlank;
+
+public record LoginResponseDto(@NotBlank String accessToken) {
+
+}
