+++ auth/src/main/java/org/bookwoori/auth/domain/member/dto/TokenRequestDto.java
@@ -1,0 +1,6 @@
+package org.bookwoori.auth.domain.member.dto;
+
+public record TokenRequestDto(
+    String refreshToken) {
+
+}
