--- auth/src/main/java/org/bookwoori/auth/dto/GetMemberResponseDto.java
@@ -1,9 +1,0 @@
-package org.bookwoori.auth.dto;
-
-public record GetMemberResponseDto(
-    Long memberId,
-    Long kakaoId,
-    String nickname,
-    String profileImg
-) {
-}
