--- auth/src/main/java/org/bookwoori/auth/dto/GetOrSaveMemberRequestDto.java
@@ -1,10 +1,0 @@
-package org.bookwoori.auth.dto;
-
-public record GetOrSaveMemberRequestDto(
-    Long kakaoId,
-    String nickname,
-    String profileImg
-) {
-
-}
-
