--- auth/src/main/java/org/bookwoori/auth/dto/LoginResponseDto.java
@@ -1,7 +1,0 @@
-package org.bookwoori.auth.dto;
-
-import jakarta.validation.constraints.NotBlank;
-
-public record LoginResponseDto(@NotBlank String accessToken) {
-
-}
