--- auth/src/main/java/org/bookwoori/auth/exception/ErrorDto.java
@@ -1,17 +1,0 @@
-package org.bookwoori.auth.exception;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-
-@Getter
-@AllArgsConstructor
-@Builder
-public class ErrorDto {
-
-    private String timestamp;
-    private int status;
-    private int code;
-    private String message;
-    private String path;
-}
