--- auth/src/main/java/org/bookwoori/auth/exception/TokenException.java
@@ -1,11 +1,0 @@
-package org.bookwoori.auth.exception;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@Getter
-@AllArgsConstructor
-public class TokenException extends RuntimeException {
-
-    private final ErrorCode errorCode;
-}
