+++ auth/src/main/java/org/bookwoori/auth/global/exception/TokenException.java
@@ -1,0 +1,11 @@
+package org.bookwoori.auth.global.exception;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@Getter
+@AllArgsConstructor
+public class TokenException extends RuntimeException {
+
+    private final ErrorCode errorCode;
+}
