+++ auth/src/main/java/org/bookwoori/auth/global/feignClient/dto/GetMemberResponseDto.java
@@ -1,0 +1,9 @@
+package org.bookwoori.auth.global.feignClient.dto;
+
+public record GetMemberResponseDto(
+    Long memberId,
+    Long kakaoId,
+    String nickname,
+    String profileImg
+) {
+}
