+++ auth/src/main/java/org/bookwoori/auth/global/feignClient/dto/GetOrSaveMemberRequestDto.java
@@ -1,0 +1,10 @@
+package org.bookwoori.auth.global.feignClient.dto;
+
+public record GetOrSaveMemberRequestDto(
+    Long kakaoId,
+    String nickname,
+    String profileImg
+) {
+
+}
+
