--- auth/src/main/java/org/bookwoori/auth/handler/CustomAuthenticationEntryPoint.java
@@ -1,49 +1,0 @@
-package org.bookwoori.auth.handler;
-
-import com.fasterxml.jackson.databind.ObjectMapper;
-import jakarta.servlet.ServletException;
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-import lombok.RequiredArgsConstructor;
-import org.bookwoori.auth.exception.ErrorCode;
-import org.bookwoori.auth.exception.ErrorDto;
-import org.springframework.security.core.AuthenticationException;
-import org.springframework.security.web.AuthenticationEntryPoint;
-import org.springframework.stereotype.Component;
-
-import java.io.IOException;
-import java.time.LocalDateTime;
-
-@Component
-@RequiredArgsConstructor
-public class CustomAuthenticationEntryPoint implements AuthenticationEntryPoint {
-
-    private final ObjectMapper objectMapper;
-
-    @Override
-    public void commence(HttpServletRequest request, HttpServletResponse response,
-        AuthenticationException authException) throws IOException, ServletException {
-        Object exception = request.getAttribute("exception");
-        if (exception instanceof ErrorCode) {
-            ErrorCode errorCode = (ErrorCode) exception;
-            setResponse(request, response, errorCode);
-            return;
-        }
-        response.sendError(HttpServletResponse.SC_UNAUTHORIZED, "Unauthorized");
-    }
-
-    private void setResponse(HttpServletRequest request, HttpServletResponse response,
-        ErrorCode errorCode) throws IOException {
-        response.setContentType("application/json;charset=UTF-8");
-        response.setStatus(HttpServletResponse.SC_UNAUTHORIZED);
-        ErrorDto errorDto = ErrorDto.builder()
-            .timestamp(LocalDateTime.now().toString())
-            .status(errorCode.getStatus())
-            .code(errorCode.getCode())
-            .message(errorCode.getMessage())
-            .path(request.getRequestURI())
-            .build();
-        String jsonString = objectMapper.writeValueAsString(errorDto);
-        response.getWriter().println(jsonString);
-    }
-}
