--- auth/src/main/java/org/bookwoori/auth/utils/CookieUtil.java
@@ -1,43 +1,0 @@
-package org.bookwoori.auth.utils;
-
-import jakarta.servlet.http.Cookie;
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-import org.springframework.stereotype.Component;
-import java.util.Arrays;
-
-@Component
-public class CookieUtil {
-
-    public static final int REFRESH_TOKEN_MAX_AGE = 7 * 24 * 60 * 60;  // 7일
-
-    public void addCookie(HttpServletResponse response, String name, String value, int maxAge) {
-        Cookie cookie = new Cookie(name, value);
-        cookie.setHttpOnly(true);
-        cookie.setSecure(true);
-        cookie.setPath("/");
-        cookie.setMaxAge(maxAge);
-        response.addCookie(cookie);
-    }
-
-    public Cookie getCookie(HttpServletRequest request, String name) {
-        if (request.getCookies() != null) {
-            return Arrays.stream(request.getCookies())
-                .filter(cookie -> name.equals(cookie.getName()))
-                .findFirst()
-                .orElse(null);
-        }
-        return null;
-    }
-
-    public void deleteCookie(HttpServletRequest request, HttpServletResponse response,
-        String name) {
-        Cookie cookie = getCookie(request, name);
-        if (cookie != null) {
-            cookie.setValue("");
-            cookie.setPath("/");
-            cookie.setMaxAge(0);
-            response.addCookie(cookie);
-        }
-    }
-}
